==> 6_3_m_e_emo_pre_loc_sleepSegments.ipynb <==
import pandas as pd
path = 'D:/DATA_THESIS/Projekt_nw_AG_AJ/Outputs_project_data_preprocessing'

e_m_day = pd.read_pickle(f'{path}/6_emotions_aggregates/m_e_emo_pre_without_sleep_impact.pkl')
location_stats = pd.read_pickle(f'{path}/5_locations/number_of_locations_for_each_day.pkl')

e_m_day.head(3)

location_stats.head(3)

location_stats['date'] = pd.to_datetime(location_stats['date'])
location_stats['dateOfSleep'] = location_stats['date'] + pd.Timedelta(days=1)
location_stats =location_stats.drop(['date'], axis=1)
location_stats.rename(columns={'count': 'count_location'}, inplace=True)

merged_df = pd.merge( e_m_day,location_stats, on=['patient_id', 'dateOfSleep'], how='inner')

merged_df

emotion = pd.read_pickle(f'{path}/6_emotions_aggregates/features_emotion.pkl')
emotion

# Merge with sleep

merged_df.head(3)

sleepSegments = pd.read_pickle(f'{path}/4_sleep_from_the_phone/all_sleepSegments.pkl')

sleepSegments

sleepSegments.rename(columns={'end_date': 'dateOfSleep'}, inplace=True)

sleepSegments = sleepSegments.drop(['hasMissingData', 'discriminator', 'start_date'], axis = 1)
sleepSegments['dateOfSleep'] = pd.to_datetime(sleepSegments['dateOfSleep'])

sleepSegments_loc_emd = pd.merge( sleepSegments,merged_df, on=['patient_id', 'dateOfSleep'], how='inner')

sleepSegments_loc_emd

sleepSegments_loc_emd.to_pickle(f'{path}/6_emotions_aggregates/m_e_emo_pre_loc_sleepSegments.pkl')
