==> 6_6_add_personality.ipynb <==
import pandas as pd

path_to_folder = 'D:/DATA_THESIS/Projekt_nw_AG_AJ/Outputs_project_data_preprocessing'


dataset = pd.read_pickle(f'{path_to_folder}/6_emotions_aggregates/m_e_emo_pre_loc_sleepSegments_steps_replaced_nan.pkl')

dataset

##  Read personalities

personality = pd.read_csv('ipip_bfm_20.csv')

personality

## Odwrócenie wartości przy pytaniach introwertycznych 

personality.columns

personality['2. Niezbyt obchodzą mnie inni ludzie.'].max()

# reverse values
for column in ['2. Niezbyt obchodzą mnie inni ludzie.',
               '3. Zostawiam moje rzeczy gdzie popadnie.',
'6. Trzymam się z boku.',
               '9. Często martwię się czymś.',
    '10. Mam trudności ze zrozumieniem abstrakcyjnych pojęć.',
       '12. Nie interesują mnie problemy innych ludzi.', 
       '13. Często zapominam odkładać rzeczy na miejsce.',
       '16. Wśród nieznajomych jestem małomówny/a.',
        '19. Często miewam huśtawki nastrojów.',
        '20. Nie mam zbyt bogatej wyobraźni.'       
]:
    personality[column] = personality[column].max() - personality[column]


personality

# Sum personality

personality['personality_sum'] = personality[[
'1. Jestem duszą towarzystwa.',
       '2. Niezbyt obchodzą mnie inni ludzie.',
       '3. Zostawiam moje rzeczy gdzie popadnie.',
       '4. Zwykle jestem zrelaksowany/a.', '5. Mam bogate słownictwo.',
       '6. Trzymam się z boku.',
       '7. Jestem wyrozumiały/a dla uczuć innych ludzi.',
       '8. Bez zwłoki wypełniam codzienne obowiązki.',
       '9. Często martwię się czymś.',
       '10. Mam trudności ze zrozumieniem abstrakcyjnych pojęć.',
       '11. Rozmawiam z wieloma różnymi ludźmi na przyjęciach.',
       '12. Nie interesują mnie problemy innych ludzi.',
       '13. Często zapominam odkładać rzeczy na miejsce.',
       '14. Rzadko czuję się przygnębiony/a.', '15. Mam głowę pełną pomysłów.',
       '16. Wśród nieznajomych jestem małomówny/a.',
       '17. Znajduję czas dla innych.',
       '18. Postępuję zgodnie z harmonogramem.',
       '19. Często miewam huśtawki nastrojów.',
       '20. Nie mam zbyt bogatej wyobraźni.'


]].sum(axis = 1)

personality

from matplotlib import pyplot as plt
plt.plot(personality['personality_sum'])

personality['patient_id'] = personality['UID']

merged_df = pd.merge(dataset, personality[['patient_id', 'personality_sum']], on='patient_id', how='inner')

merged_df

merged_df.to_pickle(f'{path_to_folder}/6_emotions_aggregates/m_e_emo_pre_loc_sleepSegments_steps_replaced_nan_personality.pkl')
